requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
plotly==5.22.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
pyyaml==6.0.1
reportlab==4.2.2
kaleido==0.2.1
flask==3.0.3
gunicorn==22.0.0
